Improve RSID Assignment (#248)
- Add a tasks to download the annovar-preprocessed version of dbsnp150
and use it to assign rsids.
- This dramatically improves the rsid match rate for the Liu et al. GWAS
of inflammatory bowel disease.
- Extend JoinDataFramesTask to allow output to Parquet. The parquet
version of this task is used for rsid assignment.
- Add a pipe to set duckdb memory limit. This was crucial to some of the
operations involved in preparing the parquet file used for rsid
assignment. In the future, I may want to chang where memory limit is
set. For example, I could set in `WF` instead of in a pipe.